Finally sort out the problem of the slow tests
